Rename V1 back inside files
