+++ libkineto/src/plugin/xpupti/XpuptiActivityProfilerSession.h
@@ -23,27 +23,24 @@
 
 namespace KINETO_NAMESPACE {
 
-std::string getXpuDeviceProperties();
+class XpuptiActivityApi;
 
 using DeviceUUIDsT = std::array<unsigned char, 16>;
 
-class XpuptiActivityApiV1;
-
-class XpuptiActivityProfilerSessionV1
+class XpuptiActivityProfilerSession
     : public libkineto::IActivityProfilerSession {
  public:
-  XpuptiActivityProfilerSessionV1() = delete;
-  XpuptiActivityProfilerSessionV1(
-      XpuptiActivityApiV1& xpti,
+  XpuptiActivityProfilerSession() = delete;
+  XpuptiActivityProfilerSession(
+      XpuptiActivityApi& xpti,
       const std::string& name,
       const libkineto::Config& config,
       const std::set<ActivityType>& activity_types);
-  XpuptiActivityProfilerSessionV1(const XpuptiActivityProfilerSessionV1&) =
-      delete;
-  XpuptiActivityProfilerSessionV1& operator=(
-      const XpuptiActivityProfilerSessionV1&) = delete;
+  XpuptiActivityProfilerSession(const XpuptiActivityProfilerSession&) = delete;
+  XpuptiActivityProfilerSession& operator=(
+      const XpuptiActivityProfilerSession&) = delete;
 
-  ~XpuptiActivityProfilerSessionV1();
+  ~XpuptiActivityProfilerSession();
 
   void start() override;
   void stop() override;
@@ -60,9 +57,7 @@ class XpuptiActivityProfilerSessionV1
   std::unique_ptr<libkineto::DeviceInfo> getDeviceInfo() override {
     return {};
   }
-  std::vector<libkineto::ResourceInfo> getResourceInfos() override {
-    return {};
-  }
+  std::vector<libkineto::ResourceInfo> getResourceInfos() override;
   std::unique_ptr<libkineto::CpuTraceBuffer> getTraceBuffer() override;
 
   void pushCorrelationId(uint64_t id) override;
@@ -109,6 +104,8 @@ class XpuptiActivityProfilerSessionV1
   // for profiling activity creation
   DeviceIndex_t getDeviceIdxFromUUID(const uint8_t deviceUUID[16]);
 
+  void addResouceInfo(int32_t device_id, int32_t sycl_queue_id);
+
  private:
   static uint32_t iterationCount_;
   static std::vector<DeviceUUIDsT> deviceUUIDs_;
@@ -125,8 +122,9 @@ class XpuptiActivityProfilerSessionV1
 
   libkineto::getLinkedActivityCallback cpuActivity_;
 
-  XpuptiActivityApiV1& xpti_;
+  XpuptiActivityApi& xpti_;
   libkineto::CpuTraceBuffer traceBuffer_;
+  std::vector<std::pair<uint32_t, uint32_t>> resourceInfo_;
   std::unordered_map<uint64_t, uint64_t> sycl_queue_pool_;
   std::unique_ptr<const libkineto::Config> config_{nullptr};
   const std::set<ActivityType>& activity_types_;
